Address review feedback: break circular dependency and fix connection error handling
REVIEW FEEDBACK ADDRESSED:
✅ **CIRCULAR DEPENDENCY RESOLVED** - moved helper functions from matrix_utils to cli_utils
✅ **CONNECTION ERROR HANDLING FIXED** - raise ConnectionError instead of silent return
✅ **ASYNCCLIENT PATCH CORRECTED** - use matrix_utils.AsyncClient for login tests

ARCHITECTURAL IMPROVEMENTS:
- **Moved _create_ssl_context()** from matrix_utils.py → cli_utils.py
- **Moved _cleanup_local_session_data()** from matrix_utils.py → cli_utils.py
- **Updated all imports** to use cli_utils for helper functions
- **Fixed test patches** to use correct module paths

TECHNICAL FIXES:
- **Connection failure handling**: main.py now raises ConnectionError for proper exit codes
- **SSL context creation**: handles certifi import gracefully
- **Test patches updated**: all _create_ssl_context and _cleanup_local_session_data patches corrected
- **Import structure cleaned**: no more circular dependency between cli_utils ↔ matrix_utils

REMAINING ISSUES:
- Login tests still failing due to complex async method mocking
- Need proper discovery_info() coroutine mocking for server discovery
- AsyncClient side_effect not working correctly for multiple client instances

Progress: Addressed all high/medium priority review feedback, continuing with test fixes.
